pyproject.toml
[tool.poetry]
name = "pythonproject3"
version = "0.1.0"
description = "first project"
authors = ["asd <pa6712@naver.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.3"
django-ninja = "^0.17.0"
isort = "^5.10.1"
black = "^22.1.0"
flake8 = "^4.0.1"
mypy = "^0.931"
django-stubs = "^1.9.0"
PyMySQL = "^1.0.2"
types-PyMySQL = "^1.0.13"
mysqlclient = "^2.1.0"
boto3 = "^1.21.21"
django-storages = "^1.12.3"
Pillow = "^9.0.1"

[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 88

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = 3.9
strict = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "app.settings"

[tool.isort]
profile = "black"